zmq_server/requirements.txt
bs4
msgpack-numpy
msgpack-python
numpy<2.0
pandas
pyqtgraph
PySide6
pyzmq
